Split Chargeback explorer into 3 explorers
Removed Accordions from Chargeback Explorer, splitting that into 3 individual explorers with their third level subtabs under Overview/Chargeback, third level tabs are Reports/Rates/Assignments. These new subtabs still use old Product features "chargeback_explorer" that will be taken care of when explorers are converted into non-explorer style screens as this will also require migration at the time.
Made changes to remove functionality from new screens to show collapsible accordion on the top like other explorer style screen
Split spec tests into multiple files based upon new controller/explorer structure

Partially fixes https://github.com/ManageIQ/manageiq-ui-classic/issues/6996
