--- app/views/chargeback/explorer.html.haml
@@ -1,11 +1,0 @@
--# These are the initial divs that will go inside center_cell_div
-#main_div
-  - case x_active_tree
-  - when :cb_reports_tree
-    = render :partial => "reports_list"
-
-  - when :cb_rates_tree
-    = render :partial => "rates_tabs"
-
-  - when :cb_assignments_tree
-    = render :partial => "assignments_tabs"
