+++ app/views/chargeback_assignment/_assignments_list.html.haml
