+++ app/views/chargeback_assignment/_cb_assignments.html.haml
