+++ app/views/chargeback_rate/_cb_rate_currency.html.haml
