+++ app/views/chargeback_rate/_cb_rate_edit.html.haml
