+++ app/views/chargeback_rate/_cb_rate_edit_table.html.haml
