+++ app/views/chargeback_rate/_cb_rate_show.html.haml
