+++ app/views/chargeback_rate/_rates_list.html.haml
