+++ app/views/chargeback_rate/_rates_tabs.html.haml
