+++ app/views/chargeback_rate/_tier_first_row.haml
