+++ app/views/chargeback_rate/_tier_row.haml
