+++ app/views/chargeback_rate/cb_rates_list.html.haml
