+++ app/views/chargeback_report/explorer.html.haml
@@ -1,0 +1,3 @@
+-# These are the initial divs that will go inside center_cell_div
+#main_div
+  = render :partial => "reports_list"
