+++ spec/controllers/chargeback_rate_controller_spec.rb
@@ -1,114 +1,5 @@
-describe ChargebackController do
+describe ChargebackRateController do
   before { stub_user(:features => :all) }
-
-  context "returns current rate assignments or set them to blank if category/tag is deleted" do
-    let(:category) { FactoryBot.create(:classification) }
-    let(:tag)      { FactoryBot.create(:classification, :parent_id => category.id) }
-    let(:entry)    { FactoryBot.create(:classification, :parent_id => tag.id) }
-
-    describe "#get_tags_all" do
-      before { entry }
-
-      it "returns the classification entry record" do
-        controller.instance_variable_set(:@edit, :cb_assign => {:tags => {}})
-        controller.send(:get_tags_all)
-
-        result = {category.id => {tag.id.to_s => tag.description}, tag.id => { entry.id.to_s => entry.description}, entry.id => {}}
-
-        expect(assigns(:edit)[:cb_assign][:tags]).to eq(result)
-      end
-    end
-
-    describe "#cb_assign_set_form_vars" do
-      before do
-        cbr = FactoryBot.create(:chargeback_rate, :rate_type => "Storage")
-        ChargebackRate.set_assignments(:Storage, [{:cb_rate => cbr, :tag => [tag, "vm"]}])
-        sandbox = {:active_tree => :cb_assignments_tree, :trees => {:cb_assignments_tree => {:active_node => 'xx-Storage'}}}
-        controller.instance_variable_set(:@sb, sandbox)
-      end
-
-      it "returns tag for current assignments" do
-        controller.send(:cb_assign_set_form_vars)
-        expect(assigns(:edit)[:current_assignment][0][:tag][0]['parent_id']).to eq(category.id)
-      end
-
-      it "returns empty array for current_assignment when tag/category is not found" do
-        tag.destroy
-        controller.send(:cb_assign_set_form_vars)
-        expect(assigns(:edit)[:current_assignment]).to eq([])
-      end
-    end
-
-    describe '#cb_assign_get_form_vars' do
-      before do
-        controller.params = {:cblabel_key => 'null'}
-        controller.instance_variable_set(:@edit, :new => {:cbshow_typ => '-labels'}, :cb_assign => {})
-      end
-
-      it "returns tag for current assignments" do
-        expect { controller.send(:cb_assign_get_form_vars) }.not_to raise_error
-      end
-
-      it "initializes hash when data are no available(params[:cblabel_key] == null)" do
-        controller.send(:cb_assign_get_form_vars)
-        docker_label_values = controller.instance_variable_get(:@edit)[:cb_assign][:docker_label_values]
-        expect(docker_label_values).to eq({})
-      end
-
-      it "initializes hash when data are no available (params[:cblabel_key] == nil)" do
-        controller.params = {:cblabel_key => nil}
-        controller.send(:cb_assign_get_form_vars)
-        docker_label_values = controller.instance_variable_get(:@edit)[:cb_assign][:docker_label_values]
-        expect(docker_label_values).to eq({})
-      end
-    end
-  end
-
-  context "Saved chargeback rendering" do
-    it "Saved chargeback reports renders paginagion buttons correctly" do
-      report = FactoryBot.create(:miq_report_with_results, :miq_group => User.current_user.current_group)
-      report.extras = {:total_html_rows => 100}
-      rp_id = report.id
-      rr_id = report.miq_report_results[0].id
-      node = "xx-#{rp_id}-2_rr-#{rr_id}"
-      html = '<table><thead><tr><th>column 1</th><th>column 2</th></thead><tbody>'
-      100.times do
-        html += '<tr><td>col1 value</td><td>col2 value</td></tr>'
-      end
-      html += '</tbody></table>'
-
-      allow(controller).to  receive(:cb_rpts_show_saved_report)
-      expect(controller).to receive(:render)
-      controller.instance_variable_set(:@sb,
-                                       :active_tree => :cb_reports_tree,
-                                       :trees       => {:cb_reports_tree => {:active_node => node}})
-      controller.instance_variable_set(:@report, report)
-      controller.instance_variable_set(:@html, html)
-      controller.instance_variable_set(:@layout, "chargeback")
-      controller.params = {:id => node}
-      controller.send(:tree_select)
-      expect(response).to                        render_template('layouts/_saved_report_paging_bar')
-      expect(controller.send(:flash_errors?)).to be_falsey
-      expect(response.status).to                 eq(200)
-    end
-
-    describe "#cb_rpt_build_folder_nodes" do
-      let!(:admin_user)        { FactoryBot.create(:user_admin) }
-      let!(:chargeback_report) { FactoryBot.create(:miq_report_chargeback_with_results) }
-
-      before { login_as admin_user }
-
-      it "returns list of saved chargeback report results" do
-        controller.send(:cb_rpt_build_folder_nodes)
-
-        parent_reports = controller.instance_variable_get(:@parent_reports)
-
-        tree_id = "#{chargeback_report.id}-0"
-        expected_result = {chargeback_report.miq_report_results.first.miq_report.name => tree_id}
-        expect(parent_reports).to eq(expected_result)
-      end
-    end
-  end
 
   describe "#explorer" do
     render_views
@@ -149,44 +40,6 @@ describe ChargebackController do
     end
   end
 
-  describe "#get_cis_all" do
-    let!(:storage) { FactoryBot.create(:storage) }
-    let!(:miq_enterprise) { FactoryBot.create(:miq_enterprise) }
-
-    it "returns names of instances of enterprise" do
-      names_miqent = {}
-      MiqEnterprise.all.each do |instance|
-        names_miqent[instance.id] = instance.name
-      end
-      controller.instance_variable_set(:@edit, :new => {:cbshow_typ => "enterprise"}, :cb_assign => {})
-      controller.send(:get_cis_all)
-      expect(assigns(:edit)[:cb_assign][:cis]).to eq(names_miqent)
-    end
-
-    it "returns names of instances of storage" do
-      names_storage = {}
-      element = "storage"
-      element.classify.constantize.all.each do |instance|
-        names_storage[instance.id] = instance.name
-      end
-      controller.instance_variable_set(:@edit, :new => {:cbshow_typ => element}, :cb_assign => {})
-      controller.send(:get_cis_all)
-      expect(assigns(:edit)[:cb_assign][:cis]).to eq(names_storage)
-    end
-
-    it "returns a ArgumentError when element not in whitelist" do
-      controller.instance_variable_set(:@edit, :new => {:cbshow_typ => "None"}, :cb_assign => {})
-      expect { controller.send(:get_cis_all) }.to raise_error(ArgumentError)
-    end
-
-    it "doesn't names of instances when nothing is selected" do
-      controller.instance_variable_set(:@edit,
-                                       :new => {:cbshow_typ => described_class::NOTHING_FORM_VALUE}, :cb_assign => {})
-      controller.send(:get_cis_all)
-      expect(assigns(:edit)[:cb_assign][:cis]).to eq({})
-    end
-  end
-
   context "chargeback rate form actions" do
     # indexing inputs regard to database
     # html inputs have names(and ids) in form "start_0_0", "finish_0_2"
@@ -217,7 +70,7 @@ describe ChargebackController do
     before do
       EvmSpecHelper.local_miq_server
       allow_any_instance_of(described_class).to receive(:center_toolbar_filename).and_return("chargeback_center_tb")
-      seed_session_trees('chargeback', :cb_rates_tree, "xx-Compute")
+      seed_session_trees('chargeback_rate', :cb_rates_tree, "xx-Compute")
       [ChargebackRateDetailMeasure, ChargeableField].each(&:seed)
     end
 
@@ -240,8 +93,8 @@ describe ChargebackController do
 
     it "renders edit form with correct values" do
       post :x_button, :params => {:pressed => "chargeback_rates_edit", :id => chargeback_rate.id}
-      expect(response).to render_template(:partial => 'chargeback/_cb_rate_edit')
-      expect(response).to render_template(:partial => 'chargeback/_cb_rate_edit_table')
+      expect(response).to render_template(:partial => 'chargeback_rate/_cb_rate_edit')
+      expect(response).to render_template(:partial => 'chargeback_rate/_cb_rate_edit_table')
 
       main_content = JSON.parse(response.body)['updatePartials']['main_div']
       expect_input(main_content, "description", "foo")
@@ -259,7 +112,7 @@ describe ChargebackController do
 
       response_body = response.body.delete('\\').gsub('u003e', ">").gsub('u003c', "<")
 
-      expect(response).to render_template(:partial => 'chargeback/_cb_rate_edit_table')
+      expect(response).to render_template(:partial => 'chargeback_rate/_cb_rate_edit_table')
 
       expect_rendered_tiers(response_body, [{:start => "0.0", :finish => "20.0"},
                                             {:start => "40.0", :finish => Float::INFINITY}])
@@ -589,36 +442,9 @@ describe ChargebackController do
     end
   end
 
-  describe "#cb_rpts_fetch_saved_report" do
-    let(:current_user) { User.current_user }
-    let(:miq_task)     { MiqTask.new(:name => "Generate Report result", :userid => current_user.userid) }
-    let(:miq_report_result) do
-      FactoryBot.create(:miq_chargeback_report_result, :miq_group => current_user.current_group, :miq_task => miq_task)
-    end
-
-    let(:chargeback_report) { FactoryBot.create(:miq_report_chargeback, :miq_report_results => [miq_report_result]) }
-
-    before do
-      miq_task.state_finished
-      miq_report_result.report = chargeback_report.to_hash.merge(:extras=> {:total_html_rows => 100})
-      miq_report_result.save
-      controller.instance_variable_set(:@sb, {})
-      controller.instance_variable_set(:@settings, :perpage => { :reports => 20 })
-    end
-
-    it "fetch existing report" do
-      controller.send(:cb_rpts_fetch_saved_report, miq_report_result.id)
-
-      fetched_report = controller.instance_variable_get(:@report)
-
-      expect(fetched_report).not_to be_nil
-      expect(fetched_report).to eq(chargeback_report)
-    end
-  end
-
   describe "#replace_right_cell" do
     it "Can build the :cb_rates tree" do
-      seed_session_trees('chargeback', :cb_rates, 'root')
+      seed_session_trees('chargeback_rate', :cb_rates, 'root')
       session_to_sb
 
       expect(controller).to receive(:render)
@@ -698,97 +524,6 @@ describe ChargebackController do
       it 'sets right cell text properly' do
         controller.send(:get_node_info, node)
         expect(controller.instance_variable_get(:@right_cell_text)).to eq("All Chargeback Rates")
-      end
-    end
-  end
-
-  context "GenericSessionMixin" do
-    let(:lastaction) { 'lastaction' }
-    let(:display) { 'display' }
-    let(:current_page) { 'current_page' }
-
-    describe '#get_session_data' do
-      it "Sets variables correctly" do
-        allow(controller).to receive(:session).and_return(:chargeback_lastaction   => lastaction,
-                                                          :chargeback_display      => display,
-                                                          :chargeback_current_page => current_page)
-        controller.send(:get_session_data)
-
-        expect(controller.instance_variable_get(:@title)).to eq("Chargeback")
-        expect(controller.instance_variable_get(:@layout)).to eq("chargeback")
-        expect(controller.instance_variable_get(:@lastaction)).to eq(lastaction)
-        expect(controller.instance_variable_get(:@display)).to eq(display)
-        expect(controller.instance_variable_get(:@current_page)).to eq(current_page)
-      end
-    end
-
-    describe '#set_session_data' do
-      it "Sets session correctly" do
-        controller.instance_variable_set(:@lastaction, lastaction)
-        controller.instance_variable_set(:@display, display)
-        controller.instance_variable_set(:@current_page, current_page)
-        controller.send(:set_session_data)
-
-        expect(controller.session[:chargeback_lastaction]).to eq(lastaction)
-        expect(controller.session[:chargeback_display]).to eq(display)
-        expect(controller.session[:chargeback_current_page]).to eq(current_page)
-      end
-    end
-  end
-
-  describe '#cb_assign_field_changed' do
-    let(:edit) do
-      {
-        :cb_assign => {
-          :cats => {
-            '1' => 'Category1',
-            '2' => 'Category2'
-          },
-          :tags => {
-            1 => {
-              '2' => 'Tag1',
-              '3' => 'Tag2',
-              '4' => 'Tag3'
-            },
-            2 => {}
-          }
-        },
-        :current   => {
-          :cbshow_typ       => 'storage-tags',
-          :cbtag_cat        => '1',
-          'storage-tags__2' => '3'
-        },
-        :new       => {
-          :cbshow_typ       => 'storage-tags',
-          :cbtag_cat        => '1',
-          'storage-tags__2' => '3'
-        }
-      }
-    end
-
-    before do
-      allow(controller).to receive(:load_edit).and_return(true)
-      controller.instance_variable_set(:@edit, edit)
-    end
-
-    context 'changing Tag Category for assignments' do
-      it 'hides buttons as no change has been made' do
-        post :cb_assign_field_changed, :params => {:cbtag_cat => '2'}
-        expect(response.body).to include("miqButtons('hide');")
-      end
-    end
-
-    context 'changing Assigned To, for assignments' do
-      it 'hides buttons as no change has been made' do
-        post :cb_assign_field_changed, :params => {:cbshow_typ => 'storage'}
-        expect(response.body).to include("miqButtons('hide');")
-      end
-    end
-
-    context 'changing item under Selections, for assignments' do
-      it 'shows buttons as a change in Selections has been made' do
-        post :cb_assign_field_changed, :params => {'storage-tags__3' => '4'}
-        expect(response.body).to include("miqButtons('show');")
       end
     end
   end
